r12490/sofa-dev : FIX: compilation with visual c++ 2005
Former-commit-id: 9913149755f40ad49f8decb629d466ea21eb34a0

+++ modules/sofa/component/projectiveconstraintset/SkeletalMotionConstraint.inl
@@ -206,7 +206,7 @@ void SkeletalMotionConstraint<DataTypes>::interpolatePosition(Real cT, typename 
                 continue;
 
             skeletonJoints[i].mLocalRigid.getCenter() = skeletonJoints[i].mPreviousMotion.getCenter() + (skeletonJoints[i].mNextMotion.getCenter() - skeletonJoints[i].mPreviousMotion.getCenter()) * dt;
-            skeletonJoints[i].mLocalRigid.getOrientation().slerp(skeletonJoints[i].mPreviousMotion.getOrientation(), skeletonJoints[i].mNextMotion.getOrientation(), (float)dt, true);
+            skeletonJoints[i].mLocalRigid.getOrientation().slerp(skeletonJoints[i].mPreviousMotion.getOrientation(), skeletonJoints[i].mNextMotion.getOrientation(), dt, true);
         }
     }
     else
